refactor(svelte2tsx): 構造化ベイクの Seg を segs.rs へ切り出す
Seg / segs_* / emit_segmented_overwrite と対応するユニットテストを
template/segs.rs へ純粋移動。出力・ロジックの変更なし。
